+++ idx/idx.go
@@ -24,6 +24,7 @@ import (
 	"sort"
 	"strings"
 	"unicode"
+	"unicode/utf8"
 
 	"golang.org/x/text/cases"
 	"golang.org/x/text/runes"
@@ -46,6 +47,69 @@ type Word struct {
 type foldedWord struct {
 	folded string
 	word   *Word
+}
+
+// whitespaceFolder will perform whitespace folding on the input. It removes
+// spaces from the beginning and end of the input and replaces all internal
+// whitespace spans with a single ASCII space rune.
+type whitespaceFolder struct {
+	// notStart is true after encounting the first non-whitespace rune.
+	notStart bool
+
+	// wsSpan is true if the transformer is currently handling a whitespace span.
+	wsSpan bool
+}
+
+// Transform implements [transform.Transformer.Transform].
+func (w *whitespaceFolder) Transform(dst, src []byte, atEOF bool) (int, int, error) {
+	var nSrc, nDst int
+	for nSrc < len(src) {
+		c, size := utf8.DecodeRune(src[nSrc:])
+		if c == utf8.RuneError && !atEOF {
+			return nDst, nSrc, transform.ErrShortSrc
+		}
+
+		isSpace := unicode.IsSpace(c)
+		if isSpace {
+			nSrc += size
+			if !w.notStart {
+				// Ignore leading whitespace.
+				continue
+			}
+			// We are in an internal whitespace span.
+			w.wsSpan = true
+			continue
+		}
+
+		if w.wsSpan {
+			// Emit a single space if we are coming out of a whitespace span.
+			// NOTE: trailing whitespace is never emitted by design.
+			spc := ' '
+			if nDst+utf8.RuneLen(spc) > len(dst) {
+				return nDst, nSrc, transform.ErrShortDst
+			}
+			nDst += utf8.EncodeRune(dst[nDst:], spc)
+			// We are no longer in an internal whitespace span.
+			w.wsSpan = false
+		}
+		w.notStart = true
+		nSrc += size
+
+		// Emit the character.
+		// NOTE: we cannot use size here because c could be utf8.RuneError in
+		// which case size would be 1 but the length of utf8.RuneError is 3.
+		if nDst+utf8.RuneLen(c) > len(dst) {
+			return nDst, nSrc, transform.ErrShortDst
+		}
+		nDst += utf8.EncodeRune(dst[nDst:], c)
+	}
+
+	return nDst, nSrc, nil
+}
+
+// Reset implements [transform.Transformer.Reset].
+func (w *whitespaceFolder) Reset() {
+	*w = whitespaceFolder{}
 }
 
 // Options are options for the idx data.
@@ -76,9 +140,20 @@ func New(r io.ReadCloser, options *Options) (*Idx, error) {
 
 	idx := &Idx{
 		foldTransformer: transform.Chain(
+			// Unicode Normalization Form D (Canonical Decomposition.
 			norm.NFD,
+			// Perform case folding.
 			cases.Fold(),
+			// Perform whitespace folding.
+			&whitespaceFolder{},
+			// Remove Non-spacing marks ([, ] {, }, etc.).
 			runes.Remove(runes.In(unicode.Mn)),
+			// Remove punctuation.
+			runes.Remove(runes.In(unicode.P)),
+			// Unicode Normalization Form C (Canonical Decomposition, followed by Canonical Composition)
+			// NOTE: Case folding does not normalize the input and may not
+			// preserve a normal form. Canonical Decomposition is thus necessary
+			// to be performed a second time.
 			norm.NFC,
 		),
 	}